requirements.txt
Pillow
exifread
requests
